Change output format with added UUIDs to metadatas
This commit discards the previous approach of writing to a jsonl file.
Instead, data is now saved as a single JSON object with two keys:
"texts" and "metadatas". Each key corresponds to a list of texts and
metadatas respectively.

Additionally, this update introduces the addition of UUIDs directly to
the metadatas within the same script.